change parameters to int

+++ src/main/java/at/tgm/ablkreim/balancer/ServerLoad.java
@@ -7,9 +7,9 @@ import java.io.Serializable;
  * @date: 23.02.2016
  */
 public class ServerLoad implements Serializable,Comparable<ServerLoad> {
-    public final double cpu, ram, drive, network;
+    public final int cpu, ram, drive, network;
 
-    public ServerLoad(double cpu, double ram, double drive, double network) {
+    public ServerLoad(int cpu, int ram, int drive, int network) {
         this.cpu = cpu;
         this.ram = ram;
         this.drive = drive;
@@ -18,6 +18,6 @@ public class ServerLoad implements Serializable,Comparable<ServerLoad> {
 
     @Override
     public int compareTo(ServerLoad other) {
-        return (int)((this.cpu + this.ram + this.drive + this.network) - (other.cpu + other.ram + other.drive + other.network));
+        return ((this.cpu + this.ram + this.drive + this.network) - (other.cpu + other.ram + other.drive + other.network));
     }
 }
